Better handling of case where model already exists
POST /v0/models/ --
Return error 400 (Bad Request) when POST request payload contains a model that already exists in the database (Meta UUID already exists)
The intended endpoint for updating a model will be PUT /v0/models

GET /v0/models/:modelId --
Previously, this returned any meta object whose UUID matched the given ID string.
Now, it ensures this meta belongs to a CDM, instead of e.g. a Diagram Element.
There will be other endpoints in the future for accessing specific Diagrams and Diagram Elements by UUID.

Startup --
On startup, the server checks to see if the example data is already in the database before adding it.

Added API.checkModelExists(modelId string) helper function for checking if a CDM is already in the database.