Abstracted model retrieval logic, WIP PUT endpoint
Made retrieveFullModel(modelId string) function, which will check to see if a model with the given ID exists, and retrieve the full thing if so.
If model not found, creates an error object and returns nil per Go common convention

Also started a (currently nonfunctional) PUT endpoint implementation. Still working this one out.