Initial changes for containerized API
Made changes to allow the API to be run in a Docker container.
Added a WIP dockerfile, which looks for **LINUX** build artifacts, copies them over, and runs them.
For this dockerfile to build successfully, you must currently:
- Build the API in a Linux environment
- Run the API once in a Linux environment to generate the database file
- Copy the built Linux binary and database file into the same directory as the dockerfile
- Build the dockerfile with something like `docker build -t opendi-api .`

To make the API container-friendly, I added CORS configuration using CORS middleware for Gin: github.com/gin-contrib/cors
Currently, the CORS config allows requests from all origins. This may be unsafe. In the future, this may be changed to the specific origin that the CDD authoring frontend will use when run locally.
For now, this setting should make the API more adaptable if anyone wishes to use it in new contexts.