Dockerfile now builds API in temp container
Improved dockerfile.
Now, the dockerfile facilitates the entire 2-stage build process for the API.
The first stage is built in an Ubuntu environment with Golang installed, and the whole Go project copied over.
Once the project is built, the Linux binary executable is copied over to the final runtime container.
This container uses a minimal distroless base debian image and the API executable only, keeping the overall image size below 40MB.

This process makes the build process fully repeatable.
The only requirement for building and deploying this API locally is the Docker engine. Containerization standardizes all the rest of it.